Apply refactor to diagnostic report
